added banner system
I want to implement a banner notification system. Banner is a personal notification for a user in the space section. A banner helps to complete neccesary actions for a participant. It's displayed at the top of the space using catch-eye colors. Banner examples:
- Fill in your Telegram Profile. It's used to communicate with the the participants by Administrator or other participants
- Fill in your GitHub profile. It's a mandatory requirement. All teams solutions produced during the hackathon must be stored in GitHub
- Find a team. Only participants embraced into teams are allowed to participate in the hackathon. Read FAQ for more information. Use a search team menu to find a vacant team. Maximum team size is 4, the minimum is 3
- For a team leader with a team less than 3 members should be shown banner with a requirement to find additional members to the team. Read FAQ for more information.

A participant might close a banner. In this case the banner is not shown anymore. The banner system should be extendable. A uniqie set of banners should be calculate every time for a user. The user might fill in the Telegram profile and the banner about it must dissapear. There are no defined rules for further banners. Likely, it will be a uniqe series of conditions in the code.

===

create a db migration for the database change.
Edit the participant profile form, rename the Github/Gitlab field to GitHub only. Mark it mandatory to fill (but in the DB it's not required to fill in)