started working on Journal
Implement Journal for Space. What is Journal? Journal is a single timeline of a participant in the hackathon. It contains record about updates, status changes, messaging, teams creating, joining approving, editing and others.

Each journal record should contain:
- Icon that defines the entity (e.g. Messaging, Profile, Team, Requests, System Events)
- Timestamp of an event
- Short description of change
- Link to the entity

The journal is personal for a participant.

A Ring bell at the right top corner must reflect new unread events in the journal.

The event marked as read when a user openned the Journal page.

The record must be tracked automaticaly from the code. I see the folowing records you must store, but if there are any you see useful, feel free to extend it:
- A Participant created
- Profile edited
- Message received
- Team created
- Team Edited
- A team Join request created
- A team join request approved or rejected
- Joined to a team
- Participant invited
- Participant joined the team
- A team member left the team
- A team ststus updated
- A team edited by the admins in dashboard
- A team deleted

The list of events is not completed and will be extended with the application update.

===

enable Journal menu item on the left section of the space