added migrations and data population for journal
Great
Now, add me a migration for the latest journal schema update. Also, add the second migration to add infromation to journal about each person profile creationg, based on the participants data. If there is an information about date and time of team creation - add it to the journal for a leader. If there is an information about date and time of a participant joining to the team - add this information to the journal. If there are no date and times on that, just create a jounral record as a fact with current date and time