added banner system
Improve the banner system. In addition to programmatic calculation of it, I want to have a frature to display custom messages from the organizers dashboard for all the participants.

Each banner must have:
- Title
- Description (HTML formatted)
- Type (info, warn)
- Date starting displaying
- Date ending displaying
- Allow to dismiss (yes | no)

The banners must be editable with the extiting editor as for messages from the dashboard menu. Add a migration to create a new entity. Run linters and build validation. Avoid adding new packages.

In the admin dashboard we should be able to add, delete, edit banner. Filter to show only active at the moment and all. By default, the banners must be shown to all the participants

===
great!\
One more thing. Make sure that on the space section maximum 2 banners are visible at one moment. If there are more notifications, collapse them and and show only 2. Add a link to expand and display all the banners.

===
Amazing!\
And let's make one more improvement to the dashboard @src/components/admin-banner-item.tsx and to space with banner display. Each banner might have an active action link. If it's presented in original banner notification, display it as an action button in the space block. Add a migration for that improvement. The links is optional for that