added calendar
Implement Calendar frature. Calendar is used to tell the participants about upcoming events. It should display all known upcoming events within a hachakon. The calendar should show a mix of events ordered by dates. A participants calendar in the space section should display a global evetns and a team-specific events.

The past events must be shown collapsed, as a past events.

Each calendar entry should contain:
- data and time of the event in the local timezone
- Event ended (optional), for long-lasting events
- title
- description (HTML formatted)
- link (optional)
- A team to which the event relates (optional)
- Active / inActive
- Type of event (info, warning, deadline), what whould be reflected in a proper colour

A banner system should reflect the closest event. The banner must be dismissable.

The calendar must be edited in the dashboard. The dashboard should be able to filter views either for all the users (global) or for a particular team. Add a button "add event" to the team page with a chance to create a specific event for this team. Use existing md-editor with conversion from MD to HTML format as it's implemented in messages and banner system. Add a migration for the database schema updates. Add event loging using a common logger. After you implemented, run linters and validations.

Make Calendar link on a sidebar active and actionable. If there are no events in the calendar view, display a text that the events will be published soon