added programming languages
I want to improve the participants and teams profiles

I want for each participant to display and edit information:
- Programming languages known and experienced in. Let's make a list of 10 the most popular languages with a chance to set personal one
- Databases known and experienced in
- Short description about person and technologies.

For a team the leader must be able to set:
- Programming language acceptable in the teams
- Stack of technology used
- Description of the team

All fields are optional to fill. Known programming languages should be seen in a public profile, and in the list of participants and teams. Don't forget to add these fields to the dashbord to edit and view teams and participants.

Add migrations for the DB

run linters and validation after you make changes.